Bug 1647816 - Make sandboxing OS version checks aware of major version. r=spohl, a=RyanVM
Fix the content and Flash sandbox policies to use the major and minor number for OS version checks allowing video playback to work properly on macOS 11 Big Sur.

Pass the macOS major and minor version to sandbox policies as a combined string.

Remove unused OSXVersion methods and members and more minor cleanup in Sandbox.mm.

Leave in the checks for unsupported versions (macOS 10.9 - 10.11) for now in case this fix needs to be uplifted.

Differential Revision: https://phabricator.services.mozilla.com/D81385