Bug 1665675 - Fix sporadic `AttributeError: module distutils has no attribute sysconfig` error in `configure` r=glandium a=RyanVM
The existing implementation of `@imports()` in the `configure` sandbox doesn't translate an import of the form `@imports('distutils.sysconfig')` into an `import distutils.sysconfig` statement; instead, it transforms the input `@imports()` request a few times in such a way that we eventually just do `import distutils`, and expect that `distutils.sysconfig` will be populated that way. This would be fine, except that this isn't the way that Python's `import` system works:

```
>>> import distutils
>>> distutils.sysconfig
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
AttributeError: module 'distutils' has no attribute 'sysconfig'
>>> import distutils.sysconfig
>>> distutils.sysconfig
<module 'distutils.sysconfig' from '/usr/lib/python3.8/distutils/sysconfig.py'>
```

i.e., we can't just import a parent module and expect that we can indirectly access all child packages of that module without importing them specifically.

So instead, we simplify the current model somewhat by not transforming the `@imports()` request at all and instead just performing the exact `import` that the user requested. This resolves the `distutils.sysconfig` issue as well as hopefully preventing any other similar issues popping up in the future.

While I'm here, I also refactored some stuff so that the way that we're patching in wrapped modules for the sandbox is more structured.

Differential Revision: https://phabricator.services.mozilla.com/D90627