Bug 1657916 - Allow extensions to see view-source requests. r=robwu, a=RyanVM
View source prefixes urls with a view-source scheme.  Somewhere along
the way a request is made for the un-prefixed url but we do not intercept.  This
removes the prefix for webrequest and proxy APIs prior to checking access rules
and filters, allowing the APIs to intercept those requests that the otherwise
could.

Differential Revision: https://phabricator.services.mozilla.com/D89068