Bug 1673870 - [devtools] Fix browser_resources_getAllResources.js intermittent. r=ladybenko, a=test-only
The test was failing because we were getting cached messages
from a worker that were emitted late.
To fix this, we create a dedicated document with workers
only when we need them and remove workers from fission_document.html.
This allows us to cleanup a few tests that were already working
around the worker logs issue.

Differential Revision: https://phabricator.services.mozilla.com/D122326