Bug 1725223 - Reset native styles of #UITourHighlightContainer properly. r=dao, a=RyanVM
This is needed for macOS / windows because of this rule:

  https://searchfox.org/mozilla-central/rev/36904ac58d2528fc59f640db57cc9429103368d3/toolkit/themes/osx/global/popup.css#12-15

This should be a better fix than double-applying the background.

Longer term it might be worth trying to make panels `appearance: none` (and no
background / borders on the panel itself, only on the contents) by default, and
add a `native` attribute or such for the ones that should look native, wdyt?

I can look into that if it'd be useful.

Differential Revision: https://phabricator.services.mozilla.com/D122583