Bug 1732577 - Don't round FT glyph extents when hinting is disabled. r=jfkthame, a=RyanVM
This bug originated from bug 1547063 when we started to allow subpixel positioning
for FreeType fonts. It would check if we requested no hinting from FreeType and
always try to round the X extents. However, light hinting disables X hinting as
well (just not Y hinting), so the behavior in this case diverged between light and
no hinting. This should make the two cases now look similar.

Differential Revision: https://phabricator.services.mozilla.com/D131537