Bug 1763634 - part 2: Make `BrowserParent` verify whether replied keyboard events are what it sent. r=smaug, a=RyanVM
For making this, `BrowserParent` (temporarily) store sent keyboard events which
wants reply from the remote process.  Then, when it gets a reply event received,
it compares whether the event data is broken or not.

The different point from the patch for mozilla-central is, there was no
`nsID::GenerateUUID` because it was introduced in 95 (bug 1723674).  Instead,
this patch uses `nsIUUIDGenerator::GenerateUUIDInPlace`.  It looks slower
since it requires to get the service.  However, it must be fine because this
is a reply for a users' key operation, i.e., not a hot path.

Differential Revision: https://phabricator.services.mozilla.com/D148242

Depends on D148241